feat(splunk): support multiple S3 log inputs
Split the Data Manager lifecycle into cohesive Python modules while preserving the Terraform entrypoint.
