+++ modules/integrations/splunk_cloud_data_manager/data_input/tests/data_input_inventory.tftest.hcl
