+++ modules/integrations/splunk_cloud_data_manager/tests/integrations_splunk_cloud_data_manager_interface_contract.tftest.hcl
@@ -73,24 +73,26 @@ run "integrations_splunk_cloud_data_manager_interface_contract" {
       "source_type             = \"\"",
       "log_group_name_prefixes = []",
       "variable \"s3_logs_config\"",
+      "s3-custom-logs = optional(list(object({",
+      "ct-logs = optional(list(object({",
+      "s3-access-logs = optional(list(object({",
+      "elb-access-logs = optional(list(object({",
+      "cf-access-logs = optional(list(object({",
       "enabled            = bool",
       "name               = string",
-      "iam_region         = optional(string, \"us-east-1\")",
+      "iam_region         = string",
       "index              = string",
-      "source_type        = string",
+      "source_type        = optional(string, \"\")",
       "sqs_urls           = list(string)",
       "s3_bucket_patterns = list(string)",
       "kms_key_arns       = list(string)",
-      "description = \"Configuration for S3 logs ingested through SQS notifications.\"",
-      "enabled            = false",
-      "name               = \"\"",
-      "index              = \"\"",
-      "source_type        = \"\"",
-      "sqs_urls           = []",
-      "s3_bucket_patterns = []",
-      "kms_key_arns       = []",
+      "description = \"Configuration for Splunk S3 log inputs grouped by dataset. Each enabled list item creates one Data Manager input.\"",
+      "default     = {}",
+      "nullable    = false",
       "validation {",
-      "condition = !var.s3_logs_config.enabled || (",
+      "Every enabled s3-custom-logs input must provide source_type.",
+      "Enabled S3 log input names must be unique across all datasets.",
+      "Enabled S3 log inputs must use unique SQS queues.",
       "variable \"default_tags\"",
       "type        = map(string)",
       "description = \"A map of tags to apply to resources.\"",
@@ -111,8 +113,8 @@ run "integrations_splunk_cloud_data_manager_interface_contract" {
       "description = \"The Splunk Cloud input map for custom logs.\"",
       "value       = var.custom_cloudwatch_log_groups_config.enabled ? local.splunk_cloud_input_custom_logs_json : \"\"",
       "output \"splunk_cloud_input_s3_logs_json\"",
-      "description = \"The Splunk Cloud input map for S3 logs.\"",
-      "value       = var.s3_logs_config.enabled ? local.splunk_cloud_input_s3_logs_json : \"\"",
+      "description = \"Map of enabled S3 log input names to their Splunk Cloud input JSON.\"",
+      "value       = tomap(local.splunk_cloud_input_s3_logs_json)",
       "output \"splunk_cloud_input_security_metadata_json\"",
       "description = \"The Splunk Cloud input map for security metadata.\"",
       "value       = var.security_metadata_config.enabled ? local.splunk_cloud_input_security_metadata_json : \"\"",
@@ -148,7 +150,7 @@ run "integrations_splunk_cloud_data_manager_interface_contract" {
     condition = (
       output.expected_input_variable_count == 10
       && output.expected_output_value_count == 4
-      && output.expected_interface_literal_count == 90
+      && output.expected_interface_literal_count == 92
     )
     error_message = "Interface contract counts must remain pinned for inputs, outputs, and source literals."
   }
