+++ modules/integrations/splunk_cloud_data_manager/tests/integrations_splunk_cloud_data_manager_source_inventory.tftest.hcl
@@ -27,9 +27,8 @@ run "integrations_splunk_cloud_data_manager_contract" {
       "local.application_config_aliases",
       "var.custom_cloudwatch_log_groups_config.enabled ? \"custom-cwl\" : \"\"",
       "var.cloudwatch_log_groups_config.enabled ? \"cwl\" : \"\"",
-      "var.s3_logs_config.enabled ? \"s3-logs\" : \"\"",
-      "splunk_s3_logs_iam_region = trimspace(var.s3_logs_config.iam_region)",
-      "provider = aws.s3_logs_iam_region",
+      "length(local.splunk_s3_logs_inputs) > 0 ? \"s3-logs\" : \"\"",
+      "region   = each.value.iam_region",
       "var.security_metadata_config.enabled ? \"secmeta\" : \"\"",
       "[\"integrations_splunk_cloud_data_manager\"]",
       "[var.aws_region]",
