+++ modules/integrations/splunk_cloud_data_manager/variables.tf
@@ -46,48 +46,135 @@ variable "custom_cloudwatch_log_groups_config" {
 
 variable "s3_logs_config" {
   type = object({
-    enabled            = bool
-    name               = string
-    iam_region         = optional(string, "us-east-1")
-    index              = string
-    source_type        = string
-    sqs_urls           = list(string)
-    s3_bucket_patterns = list(string)
-    kms_key_arns       = list(string)
+    s3-custom-logs = optional(list(object({
+      enabled            = bool
+      name               = string
+      iam_region         = string
+      index              = string
+      source_type        = optional(string, "")
+      sqs_urls           = list(string)
+      s3_bucket_patterns = list(string)
+      kms_key_arns       = list(string)
+    })), [])
+    ct-logs = optional(list(object({
+      enabled            = bool
+      name               = string
+      iam_region         = string
+      index              = string
+      sqs_urls           = list(string)
+      s3_bucket_patterns = list(string)
+      kms_key_arns       = list(string)
+    })), [])
+    s3-access-logs = optional(list(object({
+      enabled            = bool
+      name               = string
+      iam_region         = string
+      index              = string
+      sqs_urls           = list(string)
+      s3_bucket_patterns = list(string)
+      kms_key_arns       = list(string)
+    })), [])
+    elb-access-logs = optional(list(object({
+      enabled            = bool
+      name               = string
+      iam_region         = string
+      index              = string
+      sqs_urls           = list(string)
+      s3_bucket_patterns = list(string)
+      kms_key_arns       = list(string)
+    })), [])
+    cf-access-logs = optional(list(object({
+      enabled            = bool
+      name               = string
+      iam_region         = string
+      index              = string
+      sqs_urls           = list(string)
+      s3_bucket_patterns = list(string)
+      kms_key_arns       = list(string)
+    })), [])
   })
-  description = "Configuration for S3 logs ingested through SQS notifications."
-  default = {
-    enabled            = false
-    name               = ""
-    iam_region         = "us-east-1"
-    index              = ""
-    source_type        = ""
-    sqs_urls           = []
-    s3_bucket_patterns = []
-    kms_key_arns       = []
+  description = "Configuration for Splunk S3 log inputs grouped by dataset. Each enabled list item creates one Data Manager input."
+  default     = {}
+  nullable    = false
+
+  validation {
+    condition = alltrue([
+      for config in concat(
+        var.s3_logs_config["s3-custom-logs"],
+        var.s3_logs_config["ct-logs"],
+        var.s3_logs_config["s3-access-logs"],
+        var.s3_logs_config["elb-access-logs"],
+        var.s3_logs_config["cf-access-logs"],
+        ) : !config.enabled || (
+        length(trimspace(config.name)) > 0
+        && can(regex("^[a-z]{2}(-gov)?-[a-z]+-[0-9]+$", trimspace(config.iam_region)))
+        && length(trimspace(config.index)) > 0
+        && length(config.sqs_urls) > 0
+        && length(config.s3_bucket_patterns) > 0
+        && alltrue([
+          for sqs_url in config.sqs_urls :
+          can(regex("^https://sqs\\.[^.]+\\.amazonaws\\.com(\\.cn)?/[0-9]{12}/[^/?#]+$", trimspace(sqs_url)))
+        ])
+        && alltrue([
+          for value in concat(
+            config.sqs_urls,
+            config.s3_bucket_patterns,
+            config.kms_key_arns,
+          ) : length(trimspace(value)) > 0
+        ])
+      )
+    ])
+    error_message = "Every enabled S3 log input must provide a name, valid IAM roles region, non-empty index, at least one regional AWS SQS queue URL, and at least one S3 bucket pattern; configured URLs, patterns, and KMS key ARNs must not be empty."
   }
 
   validation {
-    condition = !var.s3_logs_config.enabled || (
-      length(trimspace(var.s3_logs_config.name)) > 0
-      && can(regex("^[a-z]{2}(-gov)?-[a-z]+-[0-9]+$", trimspace(var.s3_logs_config.iam_region)))
-      && length(trimspace(var.s3_logs_config.index)) > 0
-      && length(trimspace(var.s3_logs_config.source_type)) > 0
-      && length(var.s3_logs_config.sqs_urls) > 0
-      && length(var.s3_logs_config.s3_bucket_patterns) > 0
-      && alltrue([
-        for sqs_url in var.s3_logs_config.sqs_urls :
-        can(regex("^https://sqs\\.[^.]+\\.amazonaws\\.com(\\.cn)?/[0-9]{12}/[^/?#]+$", trimspace(sqs_url)))
-      ])
-      && alltrue([
-        for value in concat(
-          var.s3_logs_config.sqs_urls,
-          var.s3_logs_config.s3_bucket_patterns,
-          var.s3_logs_config.kms_key_arns,
-        ) : length(trimspace(value)) > 0
-      ])
-    )
-    error_message = "When s3_logs_config is enabled, name, a valid IAM roles region, index, source_type, at least one regional AWS SQS queue URL, and at least one S3 bucket pattern must be provided; configured URLs, patterns, and KMS key ARNs must not be empty."
+    condition = alltrue([
+      for config in var.s3_logs_config["s3-custom-logs"] :
+      !config.enabled || length(trimspace(config.source_type)) > 0
+    ])
+    error_message = "Every enabled s3-custom-logs input must provide source_type."
+  }
+
+  validation {
+    condition = length(distinct([
+      for config in concat(
+        var.s3_logs_config["s3-custom-logs"],
+        var.s3_logs_config["ct-logs"],
+        var.s3_logs_config["s3-access-logs"],
+        var.s3_logs_config["elb-access-logs"],
+        var.s3_logs_config["cf-access-logs"],
+      ) : trimspace(config.name) if config.enabled
+      ])) == length([
+      for config in concat(
+        var.s3_logs_config["s3-custom-logs"],
+        var.s3_logs_config["ct-logs"],
+        var.s3_logs_config["s3-access-logs"],
+        var.s3_logs_config["elb-access-logs"],
+        var.s3_logs_config["cf-access-logs"],
+      ) : config if config.enabled
+    ])
+    error_message = "Enabled S3 log input names must be unique across all datasets."
+  }
+
+  validation {
+    condition = length(distinct(flatten([
+      for config in concat(
+        var.s3_logs_config["s3-custom-logs"],
+        var.s3_logs_config["ct-logs"],
+        var.s3_logs_config["s3-access-logs"],
+        var.s3_logs_config["elb-access-logs"],
+        var.s3_logs_config["cf-access-logs"],
+      ) : [for sqs_url in config.sqs_urls : trimspace(sqs_url)] if config.enabled
+      ]))) == length(flatten([
+      for config in concat(
+        var.s3_logs_config["s3-custom-logs"],
+        var.s3_logs_config["ct-logs"],
+        var.s3_logs_config["s3-access-logs"],
+        var.s3_logs_config["elb-access-logs"],
+        var.s3_logs_config["cf-access-logs"],
+      ) : config.sqs_urls if config.enabled
+    ]))
+    error_message = "Enabled S3 log inputs must use unique SQS queues."
   }
 }
